restructured subviews
