+++ Scanbot.SDK.Example.Forms/Subviews/Cells/BarcodeFormatCell.cs
