+++ Scanbot.SDK.Example.Forms/Subviews/Cells/ImageResultCell.cs
